requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.5
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
gunicorn==20.0.4
psycopg2==2.8.5
SQLAlchemy==1.3.18
Flask-Session==0.3.2
Flask-Caching==1.9.0
Flask-SQLAlchemy==2.4.4
flask-redis==0.4.0
redis==3.5.3
pylibmc==1.6.1